rename btctools module to BTC
