+++ cryptotools/BTC/opcodes.py
